component split
